correct_model
sửa model
