+++ model/GioHang.js
@@ -1,9 +1,10 @@
-const db = require('./mongodbHelper');
+const db = require('../config/mongodbHelper');
 const { ObjectId } = require("mongodb");
 
 const LoaiMonSchema = new db.mongodb.Schema(
     {
-        tenLM: { type: String, default: "N/A" },
+        idKH: { type: mongoose.Schema.Types.ObjectId, ref: 'KhachHang', required: true },
+        idMon: { type: mongoose.Schema.Types.ObjectId, ref: 'Mon', required: true },
         trangThai: { type: Boolean, default: true }
     }
 );
