+++ model/HoaDon.js
@@ -1,4 +1,4 @@
-const db = require('./mongodbHelper');
+const db = require('../config/mongodbHelper');
 const { ObjectId } = require("mongodb");
 
 const HoaDonSchema = new db.mongodb.Schema(
