+++ model/LoaiMon.js
@@ -1,0 +1,13 @@
+const db = require('../config/mongodbHelper');
+const { ObjectId } = require("mongodb");
+
+const LoaiMonSchema = new db.mongodb.Schema(
+    {
+        tenLM: { type: String, default: "N/A" },
+        trangThai: { type: Boolean, default: true }
+    }
+);
+
+let model = db.mongodb.model('LoaiMon', LoaiMonSchema);
+
+module.exports = { model }
