+++ model/Mon.js
@@ -1,4 +1,4 @@
-const db = require('./mongodbHelper');
+const db = require('../config/mongodbHelper');
 const { ObjectId } = require("mongodb");
 
 const MonSchema = new db.mongodb.Schema(
